chore: moved shared files to their correct locations, fixed imports, updated vitest.config.ts to find lib shared components and hooks
