+++ frontend/src/lib/components/NotFound.tsx
